restructured libs, unique name for main script
